+++ locust/rpc/zmqrpc.py
@@ -2,28 +2,30 @@ import zmq.green as zmq
 from .protocol import Message
 
 class BaseSocket(object):
-    def __init__(self, host, port):
-        context = zmq.Context()
-        self.socket = context.socket(zmq.PAIR)
-        self._init_socket(host, port)
-
-    def _init_socket():
-        raise NotImplementedError
 
     def send(self, msg):
-        self.socket.send(msg.serialize())
+        self.sender.send(msg.serialize())
     
     def recv(self):
-        data = self.socket.recv()
+        data = self.receiver.recv()
         return Message.unserialize(data)
 
 
 class Server(BaseSocket):
-    def _init_socket(self, host, port):
-        self.socket.bind("tcp://%s:%d" % (host, port))
+    def __init__(self, host, port):
+        context = zmq.Context()
+        self.receiver = context.socket(zmq.PULL)
+        self.receiver.bind("tcp://%s:%i" % (host, port))
+        
+        self.sender = context.socket(zmq.PUSH)
+        self.sender.bind("tcp://%s:%i" % (host, port+1))
     
 
 class Client(BaseSocket):
-    def _init_socket(self, host, port):
-        self.socket.connect("tcp://%s:%d" % (host, port))
-    
+    def __init__(self, host, port):
+        context = zmq.Context()
+        self.receiver = context.socket(zmq.PULL)
+        self.receiver.connect("tcp://%s:%i" % (host, port+1))
+        
+        self.sender = context.socket(zmq.PUSH)
+        self.sender.connect("tcp://%s:%i" % (host, port))